Add files via upload
ABSTRACT
Collaborative Purchasing is a new method of buying and selling luxury assets, and any goods online.

Whilst Collaborative Purchasing has certain similarities to the elements of a traditional raffle mechanism, especially at the last stage of the process, it has a completely different business ideology of synergy between buyers and sellers. The Collaborative Purchasing platform is continuously functioning as an e-commerce e-retail venue that facilitates consumer-to-consumer and business-to-consumer sales. The core paradigm of the Collaborative Purchasing method is to provide a n innovative and powerful model/channel of e-retailing for vendors and consumers.

The Collaborative Purchasing mechanism provides easy-to-use, simple-to-understand, transparent and fair e-alternative to the on-line and off-line lotteries, and in this aspect provides distinctively higher mathematical chances of winning. The raffle-like component of the Collaborative Purchasing shall smooth the tremendous distortions and deficiencies proliferating in the contemporary lottery propositions (low chances of winning combined with occasional extraordinary large winnings) by deploying simple model and technology weaving together numerous B2C, C2C, and peer-to-peer relationships and interests, thus facilitating a continuous flow of win-win transactions.

The most appropriate immediate application of the concept is seen in purchasing luxury assets with market values exceeding the financial capacities of willing acquirers. However, the process can be equally structured to sell legitimate assets of any value.

One of the preferred embodiments of the Collaborative Purchasing method is implemented based on peer-to-peer blockchain technology.