Add files via upload
Synopsis:
TycheKo PlaySell: Collaborative Purchasing for Gamers - a New Concept of
Buying & Selling Assets, Virtual and Real
The idea shortly described in the present paper is aimed to provide easy-to-use, simple-to-understand, transparent and fair e-volutionary alternative to existing methods and means of facilitating purchasing transactions for gaming community organic to existing and perspective in- and inter-gaming environments.
PlaySell Collaborative Purchasing system has ubiquitous nature which guarantees its applicability as on-line so off-line, as for gamers so for non-gamers, which has certain similarity to closed in-community lotteries, and when compared with traditional massive lotteries and games of chance provides distinctively higher chances of winning along with establishing new powerful model/channel of e-retailing for vendors of real and virtual assets. The gaming community, who is generally distinctive by mathematical mode of thinking is, however, positioned the best to immediately appreciate the benefits of a new model.
Whilst PlaySell Collaborative Purchasing has certain similarities to group purchasing and lottery mechanisms, especially on the last stage of transactional implemention, as most picturesquely demonstrated in the blockchain version of the concept, it has completely different business ideology of synergy between buyers and sellers.
Particularly, PlaySell Collaborative Purchasing shall smooth the in- and inter-gaming economic environment by deploying simple model and technology that weaves together numerous B2C, C2C, and peer-to-peer relationships and interests, and thus facilitates a continuous flow of win-win transactions.
The most appropriate immediate application of the concept is seen in purchasing expensive items, as virtual so real/material with values exceeding the individual financial capacities of potential acquirers. However, the process can be easily streamlined to serve commercial transactions with all and any class of assets of any value, for any purchasing audience.
References to eBay.com and Amazon.com, such as samples of modified eBay.com on-line buyer interface screenshots/templates are used throughout this memo merely to simplify the presentation of the PlaySell Collaborative Purchasing principles and mechanisms by calling analogy with familiar household e-retailing platforms.
It is feasible, however, to implement the Collaborative Purchasing platform using blockchain technology.